+++ src/source/rpdb.ts
@@ -1,14 +1,23 @@
 import { ContentType } from "stremio-addon-sdk";
 import { Prefix } from "../lib/manifest.js";
-import { envGet } from "../utils/env.js";
+import { axiosHead } from "../utils/axios.js";
+import { ENV } from "../utils/env.js";
 
 const baseUrl = "https://api.ratingposterdb.com";
-export function getRpdbPoster(prefix: Prefix, id: string, type: ContentType) {
+export async function getRpdbPoster(
+  prefix: Prefix,
+  id: string,
+  type: ContentType,
+  fallbackUrl: string,
+) {
   const typeParam = type === "movie" ? "movie-" : "series-";
-  const key = envGet("RPDB_KEY") || "t0-free-rpdb";
+  const key = ENV.RPDB_API_KEY;
+  let url = fallbackUrl;
   if (prefix === Prefix.IMDB) {
-    return `${baseUrl}/${key}/imdb/poster-default/${id}?fallback=true`;
+    url = `${baseUrl}/${key}/imdb/poster-default/${id}.jpg`;
   } else {
-    return `${baseUrl}/${key}/${prefix}/poster-default/${typeParam}${id}?fallback=true`;
+    url = `${baseUrl}/${key}/${prefix}/poster-default/${typeParam}${id}.jpg`;
   }
+  if (await axiosHead<boolean>(url)) return url;
+  return fallbackUrl;
 }