+++ src/source/tmdb.ts
@@ -1,8 +1,15 @@
 import axios, { AxiosRequestConfig } from "axios";
 import { ContentType } from "stremio-addon-sdk";
 import { URLSearchParams } from "url";
-import { envGetRequired } from "../utils/env.js";
+import { ENV } from "../utils/env.js";
+import {
+  extractTitleYear,
+  matchTitle,
+  normalize,
+  Search,
+} from "../utils/fuse.js";
 import { BaseMeta, ContentDetail } from "./meta.js";
+import { Provider } from "./provider.js";
 
 export interface TmdbFindResponse {
   movie_results: TmdbMovieResult[];
@@ -30,6 +37,8 @@ export interface TmdbMovieResult {
   poster_path: string;
 }
 
+interface TmdbSeach extends Search {}
+
 interface TmdbMovieSearch {
   results: TmdbMovieResult[];
 }
@@ -38,7 +47,7 @@ interface TmdbTvSearch {
 }
 
 class TMDBService extends BaseMeta {
-  private apiKey: string = envGetRequired("TMDB_API_KEY");
+  private apiKey: string = ENV.TMDB_API_KEY;
   private baseUrl: string = "https://api.themoviedb.org/3";
   private imageUrl: string = "https://image.tmdb.org";
 
@@ -47,12 +56,16 @@ class TMDBService extends BaseMeta {
     type: ContentType,
     year?: number,
   ): Promise<ContentDetail | null> {
+    const extracted = extractTitleYear(search);
+    search = extracted.title;
+    year = year ? year : extracted.year;
     if (type === "series") {
       return await this.searchSeriesDetail(search, year);
     } else {
       return await this.searchMovieDetail(search, year);
     }
   }
+
   async findDetailImdb(
     imdbId: string,
     type: ContentType,
@@ -76,29 +89,35 @@ class TMDBService extends BaseMeta {
   }
 
   async getSearchSeries(title: string, year?: number): Promise<TmdbTvSearch> {
-    return await this._getRequest(`/search/tv?query=${title}&year=${year}`);
+    const param = { query: title, year: year };
+    return await this._getRequest(`/search/tv`, param);
   }
+
   async searchSeriesDetail(
     title: string,
     year?: number,
   ): Promise<ContentDetail | null> {
     try {
       const response = await this.getSearchSeries(title, year);
-      const tv = response.results[0];
-
-      if (tv) {
-        const year = new Date(tv.first_air_date).getFullYear();
-        this.logger.log(`Found | ${tv.name} ${year}`);
-        return {
-          id: tv.id.toString(),
-          title: tv.name,
-          overview: tv.overview,
+      const titles = response.results.map((result) => {
+        const year = new Date(result.first_air_date).getFullYear();
+        const thumbnail = result.poster_path
+          ? `${this.imageUrl}/t/p/w500${result.poster_path}`
+          : "";
+        const search: ContentDetail = {
+          id: result.id.toString(),
+          title: result.name,
+          overview: result.overview,
+          thumbnail: thumbnail,
           year: year,
           type: "movie",
-          tmdbId: tv.id,
+          tmdbId: result.id,
         };
-      }
-
+        return search;
+      });
+      const detail = matchTitle<ContentDetail>(titles, title, year);
+      const tv = detail[0];
+      if (tv) return tv;
       return null;
     } catch (error: any) {
       this.logger.error(`Movie details error | ${error.message}`);
@@ -107,29 +126,32 @@ class TMDBService extends BaseMeta {
   }
 
   async getSearchMovie(title: string, year?: number): Promise<TmdbMovieSearch> {
-    return await this._getRequest(`/search/movie?query=${title}&year=${year}`);
+    const param = { query: title, year: year };
+    return await this._getRequest(`/search/movie`, param);
   }
+
   async searchMovieDetail(
     title: string,
     year?: number,
   ): Promise<ContentDetail | null> {
     try {
       const movieResponse = await this.getSearchMovie(title, year);
-      const movie = movieResponse.results[0];
-
-      if (movie) {
+      const results = movieResponse.results.map((movie) => {
         const year = new Date(movie.release_date).getFullYear();
-        this.logger.log(`Found | ${movie.title} ${year}`);
-        return {
+        const thumbnail = `${this.imageUrl}/t/p/w500${movie.poster_path}`;
+        const search: ContentDetail = {
           id: movie.id.toString(),
           title: movie.title,
           overview: movie.overview,
+          thumbnail: thumbnail,
           year: year,
           type: "movie",
           tmdbId: movie.id,
         };
-      }
-
+        return search;
+      });
+      const movie = matchTitle(results, title, year)[0];
+      if (movie) return movie;
       return null;
     } catch (error: any) {
       this.logger.error(`Movie details error | ${error.message}`);
@@ -268,4 +290,4 @@ class TMDBService extends BaseMeta {
   }
 }
 
-export default TMDBService;
+export const tmdb = new TMDBService(Provider.TMDB);